refactor(cli): split vault_cmd.py into focused modules
Extract four modules from vault_cmd.py (2064→810 lines):
- vault_substrate.py: substrate decoders, market name lookup
- vault_rendering.py: formatting helpers (_format_amount, _print_table, etc.)
- vault_fetcher.py: on-chain data fetching, _VaultData, _safe_call
- vault_health.py: health checks, reconciliation, ERC20 balances

Click commands and orchestration stay in vault_cmd.py.
No behavior changes — pure structural refactor.
